requirements.txt
googletrans==4.0.0rc1
vncorenlp==1.0.3
pyvi
pyahocorasick
sentencepiece~=0.1.96
sacrebleu
datasets~=2.2.2
transformers~=4.17.0
sentence-transformers
pyvi
tensorboard
torch~=1.8.1
numpy==1.21
regex~=2022.3.2
python-dotenv~=0.20.0
requests~=2.27.1
tqdm~=4.62.3
setuptools~=58.0.4
flask~=1.1.2
pandas~=1.1.4
nltk~=3.5
